pyproject.toml
[build-system]
requires = ["hatchling>=1.24.0"]
build-backend = "hatchling.build"

[project]
name = "claude-orchestrator"
version = "0.1.0"
description = "Local-first orchestration system for durable Anthropic workflows."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "anthropic>=0.39.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.0",
  "python-dotenv>=1.0.0",
  "typer>=0.12.0",
  "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
claude-orchestrator = "claude_orchestrator.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["claude_orchestrator"]

[tool.hatch.build.targets.wheel.force-include]
"claude_orchestrator/migrations" = "claude_orchestrator/migrations"
"claude_orchestrator/web/templates" = "claude_orchestrator/web/templates"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

